average error fix

+++ lib/main.dart
@@ -12,7 +12,7 @@ int multiply(int a,int b) {
   return a*b;
 }
 
-int average (int a,int b,int c) {
+double average (int a,int b,int c) {
   return a+b+c/3;
 }
